Rename ManagementClient stuff to ServiceBusAdministrationClient
